requirements.txt
celery==5.4.0
pytest==8.3.3
python-dotenv==1.0.1
python-telegram-bot==21.6
redis==5.1.1
SQLAlchemy==2.0.36
flower==2.0.1
requests==2.32.3
pytest-asyncio==0.24.0